marinade_cpi

Module accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

Structs

AddLiquidity

Generated client accounts for AddLiquidity.

AddValidator

Generated client accounts for AddValidator.

ChangeAuthority

Generated client accounts for ChangeAuthority.

Claim

Generated client accounts for Claim.

ConfigMarinade

Generated client accounts for ConfigMarinade.

ConfigValidatorSystem

Generated client accounts for ConfigValidatorSystem.

DeactivateStake

Generated client accounts for DeactivateStake.

Deposit

Generated client accounts for Deposit.

DepositStakeAccount

Generated client accounts for DepositStakeAccount.

EmergencyUnstake

Generated client accounts for EmergencyUnstake.

Initialize

Generated client accounts for Initialize.

InitializeLiqPool

Generated client accounts for InitializeLiqPool.

LiquidUnstake

Generated client accounts for LiquidUnstake.

MergeStakes

Generated client accounts for MergeStakes.

OrderUnstake

Generated client accounts for OrderUnstake.

PartialUnstake

Generated client accounts for PartialUnstake.

RemoveLiquidity

Generated client accounts for RemoveLiquidity.

RemoveValidator

Generated client accounts for RemoveValidator.

SetLpParams

Generated client accounts for SetLpParams.

SetValidatorScore

Generated client accounts for SetValidatorScore.

StakeReserve

Generated client accounts for StakeReserve.

UpdateActive

Generated client accounts for UpdateActive.

UpdateActiveCommon

Generated client accounts for UpdateActiveCommon.

UpdateDeactivated

Generated client accounts for UpdateDeactivated.

UpdateDeactivatedCommon

Generated client accounts for UpdateDeactivatedCommon.