marinade_cpi

Module instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

Structs

AddLiquidity

Instruction.

AddValidator

Instruction.

ChangeAuthority

Instruction.

Claim

Instruction.

ConfigMarinade

Instruction.

ConfigValidatorSystem

Instruction.

DeactivateStake

Instruction.

Deposit

Instruction.

DepositStakeAccount

Instruction.

EmergencyUnstake

Instruction.

Initialize

Instruction.

LiquidUnstake

Instruction.

MergeStakes

Instruction.

OrderUnstake

Instruction.

PartialUnstake

Instruction.

RemoveLiquidity

Instruction.

RemoveValidator

Instruction.

SetLpParams

Instruction.

SetValidatorScore

Instruction.

StakeReserve

Instruction.

UpdateActive

Instruction.

UpdateDeactivated

Instruction.